Fix failing exception tests
