+++ hassio-google-drive-backup/backup/static/layouts/partials/error-messages.jinja2
@@ -100,7 +100,7 @@
   </ul>
 {% endcall %}
 
-{% call macros.errorMessage("Google Drive is Full", "drive_full_card", "disc_full") %}
+{% call macros.errorMessage("Google Drive is Full", "drive_full", "disc_full") %}
   <p>
     Google Drive is full, backups can't continue until space is available. You can go to Google Drive's
     <a href="https://drive.google.com/drive/quota" target="_blank" rel="noreferrer">quota page</a>
