+++ hassio-google-drive-backup/tests/conftest.py
@@ -4,6 +4,7 @@ import os
 import tempfile
 import asyncio
 import platform
+import aiohttp
 from yarl import URL
 
 import pytest
@@ -25,6 +26,7 @@ from backup.debugworker import DebugWorker
 from backup.creds import Creds
 from backup.server import ErrorStore
 from backup.ha import AddonStopper
+from backup.ui import UiServer
 from .faketime import FakeTime
 from .helpers import Uploader
 from dev.ports import Ports
@@ -58,6 +60,47 @@ class FsFaker():
         self.bytes_total = bytes_total
         if self.bytes_free > self.bytes_total:
             self.bytes_total = self.bytes_free
+
+
+class ReaderHelper:
+    def __init__(self, session, ui_port, ingress_port):
+        self.session = session
+        self.ui_port = ui_port
+        self.ingress_port = ingress_port
+        self.timeout = aiohttp.ClientTimeout(total=20)
+
+    def getUrl(self, ingress=True, ssl=False):
+        if ssl:
+            protocol = "https"
+        else:
+            protocol = "http"
+        if ingress:
+            return protocol + "://localhost:" + str(self.ingress_port) + "/"
+        else:
+            return protocol + "://localhost:" + str(self.ui_port) + "/"
+
+    async def getjson(self, path, status=200, json=None, auth=None, ingress=True, ssl=False, sslcontext=None):
+        async with self.session.get(self.getUrl(ingress, ssl) + path, json=json, auth=auth, ssl=sslcontext, timeout=self.timeout) as resp:
+            assert resp.status == status
+            return await resp.json()
+
+    async def get(self, path, status=200, json=None, auth=None, ingress=True, ssl=False):
+        async with self.session.get(self.getUrl(ingress, ssl) + path, json=json, auth=auth, timeout=self.timeout) as resp:
+            if resp.status != status:
+                import logging
+                logging.getLogger().error(resp.text())
+                assert resp.status == status
+            return await resp.text()
+
+    async def postjson(self, path, status=200, json=None, ingress=True):
+        async with self.session.post(self.getUrl(ingress) + path, json=json, timeout=self.timeout) as resp:
+            assert resp.status == status
+            return await resp.json()
+
+    async def assertError(self, path, error_type="generic_error", status=500, ingress=True, json=None):
+        logging.getLogger().info("Requesting " + path)
+        data = await self.getjson(path, status=status, ingress=ingress, json=json)
+        assert data['error_type'] == error_type
 
 
 # This module should onyl ever have bindings that can also be satisfied by MainModule
@@ -125,6 +168,20 @@ async def injector(cleandir, server_url, ports):
 
     # logging.getLogger('injector').setLevel(logging.DEBUG)
     return Injector([BaseModule(config), TestModule(ports)])
+
+
+@pytest.fixture
+async def ui_server(injector, server):
+    os.mkdir("static")
+    server = injector.get(UiServer)
+    await server.run()
+    yield server
+    await server.shutdown()
+
+
+@pytest.fixture
+def reader(ui_server, session, ui_port, ingress_port):
+    return ReaderHelper(session, ui_port, ingress_port)
 
 
 @pytest.fixture
